feat : add permission verifiications

+++ src/main/java/africa/norsys/doc/controller/DocumentController.java
@@ -1,6 +1,7 @@
 package africa.norsys.doc.controller;
 
 import africa.norsys.doc.entity.Document;
+import africa.norsys.doc.enumerator.Permission;
 import africa.norsys.doc.exception.DocumentNotAddedException;
 import africa.norsys.doc.exception.DocumentNotFoundException;
 import africa.norsys.doc.service.DocumentService;
@@ -41,7 +42,16 @@ public class DocumentController {
 
 
     @GetMapping("/file/{filename:.+}")
-    public ResponseEntity<byte[]> getFile(@PathVariable String filename) throws IOException {
+    public ResponseEntity<byte[]> getFile(@PathVariable String filename,
+                                          @RequestParam UUID documentId,
+                                          @RequestParam("userId") UUID userId) throws IOException {
+        // Check if the user has access to the document
+        boolean hasAccess = documentService.checkUserAccess(documentId, userId, Permission.READ);
+
+        if (!hasAccess) {
+            // If the user does not have access, return a forbidden status code
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         byte[] fileBytes = documentService.getFileBytes(filename);
         return ResponseEntity.ok().body(fileBytes);
     }
@@ -59,8 +69,18 @@ public class DocumentController {
     }
 
     @DeleteMapping("/{documentId}")
-    public ResponseEntity<Void> deleteDocument(@PathVariable UUID documentId) {
+    public ResponseEntity<Void> deleteDocument(@PathVariable UUID documentId, @RequestParam("userId") UUID userId) {
         try {
+
+            // Check if the user has access to the document
+            boolean hasAccess = documentService.checkUserAccess(documentId, userId, Permission.WRITE);
+
+            if (!hasAccess) {
+                // If the user does not have access, return a forbidden status code
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+
+
             documentService.deleteDocumentById(documentId);
             return ResponseEntity.ok().build();
         } catch (DocumentNotFoundException e) {
@@ -70,12 +90,24 @@ public class DocumentController {
         }
     }
 
-    @GetMapping("/{id}")
-    public ResponseEntity<Document> getDocumentById(@PathVariable UUID id) {
-        Optional<Document> optionalDocument = documentService.getDocumentById(id);
+    @GetMapping("/{documentId}")
+    public ResponseEntity<Document> getDocumentById(
+            @PathVariable UUID documentId,
+            @RequestParam("userId") UUID userId) {
+        // Check if the user has access to the document
+        boolean hasAccess = documentService.checkUserAccess(documentId, userId, Permission.READ);
+
+        if (!hasAccess) {
+            // If the user does not have access, return a forbidden status code
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        // If the user has access, retrieve and return the document details
+        Optional<Document> optionalDocument = documentService.getDocumentById(documentId);
         return optionalDocument.map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
     }
+
 
     @GetMapping("/search")
     public ResponseEntity<Page<Document>> searchDocuments(
